// fe/src/lib/api/generated/model/maintenanceResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface MaintenanceResponse {
  manual_expected_end_at: string | null;
  message: string;
  schedule_active: boolean;
  schedule_upcoming: boolean;
  scheduled_end_at: string | null;
  scheduled_start_at: string | null;
  source: string;
  state: string;
}